[ALPS07168164] apusys kernel midware: fix coverity
[Detail]
Fix coverity

MTK-Commit-Id: b68d84dac6e807bff00d1304a7cd2875019690f6

CR-Id: ALPS07168164
Feature: apu_sys_kernel

+++ drivers/misc/mediatek/apusys/midware/2.0/rv/mdw_rv_events.h
@@ -46,7 +46,7 @@ TRACE_EVENT(mdw_rv_cmd,
 		__field(uint32_t, priority)
 		__field(uint32_t, softlimit)
 		__field(uint32_t, pwr_dtime)
-		__field(uint32_t, sc_rets)
+		__field(uint64_t, sc_rets)
 	),
 	TP_fast_assign(
 		__entry->done = done;
